requirements.txt
mediapipe==0.10.9
cvzone==1.6.1
opencv-python==4.9.0
PyYAML==6.0.1